+++ integrations/va-okta/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gitbook/integration-va-okta
 
+## 0.0.9
+
+### Patch Changes
+
+-   6181f25: Move VA integrations to their own category
+
+## 0.0.8
+
+### Patch Changes
+
+-   15a814a: Add site:metadata:read scope to VA integrations
+
 ## 0.0.7
 
 ### Patch Changes