+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @gitbook/cli
 
+## 0.17.0
+
+### Minor Changes
+
+-   d3fcacc: Remove entities from CLI
+
+### Patch Changes
+
+-   Updated dependencies [d3fcacc]
+    -   @gitbook/api@0.57.0
+
+## 0.16.0
+
+### Minor Changes
+
+-   6646aac: Update the CLI to accommodate new category
+
 ## 0.15.2
 
 ### Patch Changes